requirements.txt
dj-database-url==0.5.0
Django==2.1.2
djangorestframework==3.9.0
python-decouple==3.1
tox==3.5.2
